Added props
I added a silver/gold key, a coin, the weapon switch ui, and, at least
for the first level, only two skull change things. Basically, it starts
off as a white skull and after the player kills someone, the skull is
half filled to represent one kill. after the player kills the second
person, the skull is fully filled. This is just a working concept
though. not set in stone